first 50 fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,7 +12,7 @@ int main(void)
 {
 	int a = 0;
 	int b = 1;
-	int next = a + b;
+	long double next = a + b;
 	int i;
 
 	printf("%d, %d, ", a, b);
